add names package (#6)
Ports the `github.com/aws-controllers-k8s/code-generator/pkg/names`
package to the new repository, making one change to the package:
adding a `Names.SnakeStripped` field that contains the snake-cased
name stripped of any non-alphanumeric characters. This is useful for
determining the Go package name from a resource name...

Signed-off-by: Jay Pipes <jaypipes@gmail.com>

Issue #, if available:

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.