Introduce `scalar` package
This patch introduces a new Go package named `scalar`, which aims to
group utility functions that are commonly used accross controllers,
code-generator and runtime. These functions operate on primitive types
like strings, int, bool, etc... The package mainly brings two new
functions:

- `EqualPStrict`: returns true if two given scalar pointers are
strictly equal, meaning that they either have the same nility or the
same value.
- `EqualP`: returns true if two given scalars represent the same value.
It behaves similarly to `EqualPStrict` but will also return true if one
of the parameters is nil and the other points to a zero scalar value.

In addition, this patch removes the `strutil` package that used to have
utility functions `InStrings` and `InStringsPtr`. Thos functions are now
available under the names `Contains` and `ContainsP` and can be reused
will the other primitive types as well.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>