Handle special cases for IPv4 and IPv6 (#29)
This commit addresses the issue of `IPv4` and `IPv6` being incorrectly
converted to `iPv4` and `iPv6` respectively during the lower-casing
process. It adds new entries to the initialisms variable to ensure that
`IPv4` and `IPv6` are correctly preserved during the conversion phase.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.